YUR-99: Fixed YUR-99 PR comments.

+++ src/main/java/com/yurjinia/project_structure/column/dto/UpdateColumnPositionRequest.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class UpdateColumnPositionRequest {
     private String columnName;
-    private long columnPosition;
+    private int columnPosition;
 }
